Move updater to separate app
Mainly in an attempt to make it actually start up, although that was a
red herring. Fixed it so the app doesn't crash if the HEAD request to
check the new firmware fails – I think it was crashing so quickly on
startup because httpotion/poison weren't ready that the supervisor gave
up restarting it.